Refactor and clean up floodfill code
No functional changes, with the exception of
removed SWIG directives.

New separate header for tile constants.

Morph, blur and gap detection split up into separate files.

Separate header files for SWIG-generation of interfaces for
the morph/blur/gap detection structures and functions used
from python.

Commenting style made more consistent.
