+++ lib/fill_constants.hpp
@@ -1,0 +1,31 @@
+/* This file is part of MyPaint.
+ * Copyright (C) 2019 by the MyPaint Development Team.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ */
+
+#ifndef FILL_CONSTANTS_HPP
+#define FILL_CONSTANTS_HPP
+
+#include "fill_common.hpp"
+
+/*
+  Constant tiles, mostly used for pointer comparison
+  but also read from into nine-grid data sections.
+*/
+class ConstTiles
+{
+  public:
+    static PyObject* ALPHA_OPAQUE();
+    static PyObject* ALPHA_TRANSPARENT();
+
+  private:
+    static void init();
+    static PyObject* _ALPHA_OPAQUE;
+    static PyObject* _ALPHA_TRANSPARENT;
+};
+
+#endif //FILL_CONSTANTS_HPP
