+++ lib/morphology.cpp
@@ -6,10 +6,11 @@
  * the Free Software Foundation; either version 2 of the License, or
  * (at your option) any later version.
  */
+
 #include "morphology.hpp"
+#include "fill_constants.hpp"
+
 #include <cmath>
-#include <thread>
-#include <future>
 #include <tuple>
 
 MorphBucket::MorphBucket(int radius)
@@ -278,47 +279,6 @@ erode(MorphBucket& mb, bool update_input, bool update_lut, GridVector input)
         mb, update_input, update_lut, input);
 }
 
-// For the given tile coordinate, return a vector of pixel buffers for
-// the tiles of the coordinate and its 8 neighbours. If a neighbouring
-// tile does not exist, the empty alpha tile takes its place.
-// Order of tiles in vector, where 4 is the input tile:
-// 0 1 2
-// 3 4 5
-// 6 7 8
-GridVector
-nine_grid(PyObject* tile_coord, PyObject* tiles)
-{
-    const int num_tiles = 9;
-    const int offs[]{-1, 0, 1};
-
-    int x, y;
-
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-
-    PyArg_ParseTuple(tile_coord, "ii", &x, &y);
-    std::vector<PixelBuffer<chan_t>> grid;
-
-    for (int i = 0; i < num_tiles; ++i) {
-        int _x = x + offs[i % 3];
-        int _y = y + offs[i / 3];
-        PyObject* c = Py_BuildValue("ii", _x, _y);
-        PyObject* tile = PyDict_GetItem(tiles, c);
-        Py_DECREF(c);
-        if (tile)
-            grid.push_back(PixelBuffer<chan_t>(tile));
-        else
-            grid.push_back(
-                PixelBuffer<chan_t>(ConstTiles::ALPHA_TRANSPARENT()));
-    }
-    PyGILState_Release(gstate);
-
-    return grid;
-}
-
-using Strand = AtomicQueue<PyObject*>;
-using StrandQueue = AtomicQueue<Strand>;
-
 // Morph a single strand of tiles, storing
 // the output tiles in a Python dictionary "morphed"
 void
@@ -354,10 +314,6 @@ morph_strand(
     }
 }
 
-// Worker, processing strands of tiles from a distributed workload
-using worker_function =
-    std::function<void(int, StrandQueue&, PyObject*, std::promise<PyObject*>)>;
-
 void
 morph_worker(
     int offset, StrandQueue& queue, PyObject* tiles,
@@ -375,55 +331,6 @@ morph_worker(
     result.set_value(morphed);
 }
 
-static int
-num_strand_workers(int num_strands, int min_strands_per_worker)
-{
-    int max_threads = std::thread::hardware_concurrency();
-    int max_by_strands = num_strands / min_strands_per_worker;
-    return MAX(1, MIN(max_threads, max_by_strands));
-}
-
-void
-process_strands(
-    worker_function worker, int offset, int min_strands_per_worker,
-    PyObject* strands, PyObject* tiles, PyObject* morphed)
-{
-    Py_ssize_t num_strands = PyList_GET_SIZE(strands);
-    int num_threads = num_strand_workers(num_strands, min_strands_per_worker);
-
-    std::vector<std::thread> threads(num_threads);
-    std::vector<std::future<PyObject*>> futures(num_threads);
-
-    PyEval_InitThreads();
-    StrandQueue work_queue(strands);
-
-    // Create worker threads
-    for (int i = 0; i < num_threads; ++i) {
-        std::promise<PyObject*> promise;
-        futures[i] = promise.get_future();
-        threads[i] = std::thread(
-            worker, offset, std::ref(work_queue), tiles, std::move(promise));
-    }
-    // Release the lock to let the workers work
-    Py_BEGIN_ALLOW_THREADS
-
-    for (int i = 0; i < num_threads; ++i)
-    {
-        // Wait for the output from the threads
-        // and merge it into the final result
-        futures[i].wait();
-        PyObject* _m = futures[i].get();
-        PyGILState_STATE state;
-        state = PyGILState_Ensure();
-        PyDict_Update(morphed, _m);
-        Py_DECREF(_m);
-        PyGILState_Release(state);
-        threads[i].join();
-    }
-
-    Py_END_ALLOW_THREADS
-}
-
 // Entry point to morphological operations,
 // this is what should be called from Python code.
 void
@@ -439,139 +346,6 @@ morph(int offset, PyObject* morphed, PyObject* tiles, PyObject* strands)
         morph_worker, offset, min_strands_per_worker, strands, tiles, morphed);
 }
 
-// Box blur parameters & memory allocation
-
-// Generate gaussian multiplicands used for blurring.
-// They are stored and used with fixed-point arithmetic
-static const std::vector<fix15_short_t>
-blur_factors(int r)
-{
-    constexpr double pi = 3.141592653589793;
-
-    // Equations nicked from Krita
-    float sigma = 0.3 * r + 0.3;
-    int prelim_size = 6 * std::ceil(sigma + 1);
-    float mul = 1 / sqrt(2 * pi * sigma * sigma);
-    float exp_mul = 1 / (2 * sigma * sigma);
-
-    std::vector<fix15_short_t> factors;
-    int center = (prelim_size - 1) / 2;
-    for (int i = 0; i < prelim_size; ++i) {
-        int d = center - i;
-        double fac = mul * exp(-d * d * exp_mul);
-        // The bit-or'ing is a hack to avoid the sum of
-        // multiplicands being less than 0, blurred pixels
-        // are clamped to fix15_one anyway.
-        factors.push_back((fix15_t)(fix15_one * fac) | 3);
-    }
-    return factors;
-}
-
-// Allocate memory for input and intermediate buffers
-BlurBucket::BlurBucket(int r)
-    : factors(blur_factors(r)), radius((factors.size() - 1) / 2)
-{
-    // Suppress uninitialization warning, the output
-    // array is always fully populated before use
-    output[0][0] = 0;
-    const int d = N + radius * 2;
-    // Output from 3x3-grid,
-    // input to horizontal blur (Y x X) = (d x d)
-    input_full = new chan_t*[d];
-    for (int i = 0; i < d; ++i) {
-        input_full[i] = new chan_t[d];
-    }
-    // Output for horizontal blur,
-    // input to vertical blur (Y x X) = (d x N)
-    input_vert = new chan_t*[d];
-    for (int i = 0; i < d; ++i) {
-        input_vert[i] = new chan_t[N];
-    }
-}
-
-BlurBucket::~BlurBucket()
-{
-    const int d = N + radius * 2;
-    for (int i = 0; i < d; ++i) {
-        delete[] input_full[i];
-        delete[] input_vert[i];
-    }
-    delete[] input_full;
-    delete[] input_vert;
-}
-
-PyObject*
-BlurBucket::blur(bool can_update, GridVector input_grid)
-{
-    initiate(can_update, input_grid);
-
-    if (input_fully_opaque()) return ConstTiles::ALPHA_OPAQUE();
-
-    if (input_fully_transparent()) return ConstTiles::ALPHA_TRANSPARENT();
-
-    int r = radius;
-
-    // Blur each row from input to intermediate buffer
-    for (int y = 0; y < N + 2 * r; ++y) {
-        for (int x = 0; x < N; ++x) {
-            fix15_t blurred = 0;
-            for (int xoffs = -r; xoffs < r + 1; xoffs++) {
-                fix15_t in = input_full[y][x + xoffs + r];
-                blurred += fix15_mul(in, factors[xoffs + r]);
-            }
-            input_vert[y][x] = fix15_short_clamp(blurred);
-        }
-    }
-
-    // Blur each column from intermediate to output buffer
-    for (int x = 0; x < N; ++x) {
-        for (int y = 0; y < N; ++y) {
-            fix15_t blurred = 0;
-            for (int yoffs = -r; yoffs < r + 1; yoffs++) {
-                fix15_t in = input_vert[y + yoffs + r][x];
-                blurred += fix15_mul(in, factors[yoffs + r]);
-            }
-            output[y][x] = fix15_short_clamp(blurred);
-        }
-    }
-
-    npy_intp dims[] = {N, N};
-
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-
-    PyObject* out_tile = PyArray_EMPTY(2, dims, NPY_USHORT, 0);
-    PixelBuffer<chan_t> out_buf(out_tile);
-
-    PyGILState_Release(gstate);
-
-    PixelRef<chan_t> out_px = out_buf.get_pixel(0, 0);
-    for (int y = 0; y < N; ++y) {
-        for (int x = 0; x < N; ++x) {
-            out_px.write(output[y][x]);
-            out_px.move_x(1);
-        }
-    }
-
-    return out_tile;
-}
-
-void
-BlurBucket::initiate(bool can_update, GridVector input)
-{
-    init_from_nine_grid(radius, input_full, can_update, input);
-}
-
-template <typename T>
-static bool
-all_eq(T** arr, int dim, T val)
-{
-    for (int y = 0; y < dim; ++y)
-        for (int x = 0; x < dim; ++x)
-            if (arr[y][x] != val) return false;
-    return true;
-}
-
 bool
 MorphBucket::input_fully_opaque()
 {
@@ -582,225 +356,4 @@ bool
 MorphBucket::input_fully_transparent()
 {
     return all_eq<chan_t>(input, 2 * radius + N, 0);
-}
-
-bool
-BlurBucket::input_fully_opaque()
-{
-    return all_eq<chan_t>(input_full, 2 * radius + N, fix15_one);
-}
-
-bool
-BlurBucket::input_fully_transparent()
-{
-    return all_eq<chan_t>(input_full, 2 * radius + N, 0);
-}
-
-/*
-  Blur a strand of tiles, from top to bottom. This function
-  is very similar to morph_strand, but does not need to keep
-  track of separate update flags. In fact, since the blur
-  function always writes its input array before checking if it
-  can skip the actual op, subsequent blurs can always update
-  the input array.
-*/
-void
-blur_strand(
-    Strand& strand, PyObject* tiles, BlurBucket& bucket, PyObject* blurred)
-{
-    bool can_update = false;
-    PyObject* tile_coord;
-    while (strand.pop(tile_coord)) {
-        GridVector grid = nine_grid(tile_coord, tiles);
-
-        PyObject* result = bucket.blur(can_update, grid);
-        can_update = true;
-
-        // Add morphed tile unless it is completely transparent
-        bool is_empty = result == ConstTiles::ALPHA_TRANSPARENT();
-        bool is_full = result == ConstTiles::ALPHA_OPAQUE();
-        PyGILState_STATE gstate;
-        gstate = PyGILState_Ensure();
-        if (!is_empty) PyDict_SetItem(blurred, tile_coord, result);
-        if (!(is_empty || is_full)) Py_DECREF(result);
-        PyGILState_Release(gstate);
-    }
-}
-
-void
-blur_worker(
-    int radius, StrandQueue& queue, PyObject* tiles,
-    std::promise<PyObject*> result)
-{
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-    PyObject* blurred = PyDict_New();
-    PyGILState_Release(gstate);
-    BlurBucket bucket(radius);
-    Strand strand;
-    while (queue.pop(strand)) {
-        blur_strand(strand, tiles, bucket, blurred);
-    }
-    result.set_value(blurred);
-}
-
-void
-blur(int radius, PyObject* blurred, PyObject* tiles, PyObject* strands)
-{
-    if (radius <= 0 || !PyDict_Check(tiles) || !PyList_CheckExact(strands)) {
-        printf("Invalid blur parameters!\n");
-        return;
-    }
-
-    const int min_strands_per_worker = 2;
-    process_strands(
-        blur_worker, radius, min_strands_per_worker, strands, tiles, blurred);
-}
-
-// Gap closing stuff - prob move this out somewhere
-
-DistanceBucket::DistanceBucket(int distance) : distance(distance)
-{
-    int r = N + distance * 2 + 2;
-    input = new chan_t*[r];
-    for (int i = 0; i < r; ++i)
-        input[i] = new chan_t[r];
-}
-
-DistanceBucket::~DistanceBucket()
-{
-    int r = N + distance * 2 + 2;
-    for (int i = 0; i < r; ++i)
-        delete[] input[i];
-    delete[] input;
-}
-
-typedef coord (*rot_op)(int x, int y, int xoff, int yoff);
-
-static inline void
-upd_dist(coord lc, PixelBuffer<chan_t>& dists, int new_dst)
-{
-    if (lc.x < 0 || lc.x > (N - 1) || lc.y < 0 || lc.y > (N - 1)) return;
-    int curr_dist = dists(lc.x, lc.y);
-    if (curr_dist > new_dst) {
-        dists(lc.x, lc.y) = new_dst;
-    }
-}
-
-// Search an octant with a radius of _dist_ pixels, marking any gaps
-// that are found. The octant searched is determined by the rotation
-// function provided.
-bool
-dist_search(
-    int x, int y, int dist, chan_t** alphas, PixelBuffer<chan_t>& dists,
-    rot_op op)
-{
-
-    // int d_lim = 1 + (dist * dist);
-    int offs = dist + 1;
-    int rx = x - offs;
-    int ry = y - offs;
-
-    coord t1 = op(x, y, 0, -1);
-    coord t2 = op(x, y, 1, -1);
-
-    if (alphas[t1.y][t1.x] == 0 || alphas[t2.y][t2.x] == 0) return false;
-
-    bool gap_found = false;
-
-    for (int yoffs = 2; yoffs < dist + 2; ++yoffs) {
-        int y_dst_sqr = (yoffs - 1) * (yoffs - 1);
-
-        for (int xoffs = 0; xoffs <= yoffs; ++xoffs) {
-            int offs_dst = y_dst_sqr + (xoffs) * (xoffs);
-            // int real_offs = (int) round(sqrtf((float) offs_dst));
-            if (offs_dst >= 1 + dist * dist) break;
-            coord c = op(x, y, xoffs, -yoffs);
-            if (alphas[c.y][c.x] == 0) { // Gap found
-
-                gap_found = true;
-
-                // Double-width distance assignment
-                float dx = (float)xoffs / (yoffs - 1);
-                float tx = 0;
-                int cx = 0;
-                for (int cy = 1; cy < yoffs; ++cy) {
-                    upd_dist(op(rx, ry, cx, 0 - cy), dists, offs_dst);
-                    tx += dx;
-                    if (floor(tx) > cx) {
-                        cx++;
-                        upd_dist(op(rx, ry, cx, 0 - cy), dists, offs_dst);
-                    }
-                    upd_dist(op(rx, ry, cx + 1, 0 - cy), dists, offs_dst);
-                }
-            }
-        }
-    }
-    return gap_found;
-}
-
-// Coordinate reflection/rotation
-coord
-top_right(int x, int y, int xoffs, int yoffs)
-{
-    return coord(x + xoffs, y + yoffs);
-}
-coord
-top_centr(int x, int y, int xoffs, int yoffs)
-{
-    return coord(x - yoffs, y - xoffs);
-}
-coord
-bot_centr(int x, int y, int xoffs, int yoffs)
-{
-    return coord(x - yoffs, y + xoffs);
-}
-coord
-bot_right(int x, int y, int xoffs, int yoffs)
-{
-    return coord(x + xoffs, y - yoffs);
-}
-
-/* Search for gaps in the 9-grid of flooded alpha tiles,
-   a gap being defined as a
- */
-bool
-find_gaps(
-    DistanceBucket& rb, PyObject* radiuses_arr, PyObject* mid, PyObject* n,
-    PyObject* e, PyObject* s, PyObject* w, PyObject* ne, PyObject* se,
-    PyObject* sw, PyObject* nw)
-{
-    int r = rb.distance + 1;
-
-    typedef PixelBuffer<chan_t> PBT;
-    GridVector input{PBT(nw), PBT(n),  PBT(ne), PBT(w), PBT(mid),
-                     PBT(e),  PBT(se), PBT(s),  PBT(sw)};
-
-    init_from_nine_grid(r, rb.input, false, input);
-
-    bool gaps_found = false;
-
-    PixelBuffer<chan_t> radiuses(radiuses_arr);
-    // search for gaps in an approximate semi-circle
-    for (int y = 0; y < 2 * r + N - 1;
-         ++y) { // we check at most distance+1 pixels above any point
-        for (int x = 0; x < r + N - 1; ++x) {
-            if (rb.input[y][x] ==
-                0) { // Search for gaps in relation to this pixel
-                if (y >= r) {
-                    gaps_found |= dist_search(
-                        x, y, rb.distance, rb.input, radiuses, top_right);
-                    gaps_found |= dist_search(
-                        x, y, rb.distance, rb.input, radiuses, top_centr);
-                }
-                if (y < N + r) {
-                    gaps_found |= dist_search(
-                        x, y, rb.distance, rb.input, radiuses, bot_centr);
-                    gaps_found |= dist_search(
-                        x, y, rb.distance, rb.input, radiuses, bot_right);
-                }
-            }
-        }
-    }
-    return gaps_found;
 }
