+++ lib/morphology.hpp
@@ -7,12 +7,11 @@
  * (at your option) any later version.
  */
 
-#ifndef MORPH_HPP
-#define MORPH_HPP
+#ifndef MORPHOLOGY_HPP
+#define MORPHOLOGY_HPP
 
 #include "fill_common.hpp"
-
-#include <vector>
+#include "morphology_swig.hpp"
 
 /*
   Chords make up the structuring elements used to perform morphological
@@ -23,9 +22,6 @@
   is stored, avoiding redundant lookups/tests for chords of the same
   lengths..
 */
-#ifdef SWIG
-%ignore chord;
-#endif
 struct chord {
     chord() : x_offset(0), length_index(0) {}
     chord(int x, int len_i) : x_offset(x), length_index(len_i){};
@@ -50,9 +46,6 @@ typedef chan_t op(chan_t, chan_t);
   Output array to store morphed alpha values (consider removing/replacing).
 */
 
-#ifdef SWIG
-%ignore MorphBucket;
-#endif
 class MorphBucket
 {
   public:
@@ -79,74 +72,4 @@ class MorphBucket
     chan_t** input; // input 2d array populated by 3x3 input tile grid
 };
 
-// Perform a dilation or erosion using the given input tiles
-// and strands of vertically contiguous coordinates, placing
-// the result in the given coord->tile dictionary.
-void morph(
-    int offset, // Radius to grow (if > 0) or shrink (if < 0)
-    PyObject* morphed, // Dictionary holding the result of the operation
-    PyObject* tiles, // Input tiles, NxNx1 uint16 numpy arrays
-    PyObject* strands // Strands of contiguous tile coordinates
-    );
-
-#ifdef SWIG
-%ignore BlurBucket;
-#endif
-
-/*
-  Holds data and allocated space used to perform
-  tile-wise box blur.
-*/
-
-class BlurBucket
-{
-  public:
-    explicit BlurBucket(int radius);
-    ~BlurBucket();
-    PyObject* blur(bool can_update, GridVector input);
-
-  private:
-    void initiate(bool can_update, GridVector input);
-    bool input_fully_opaque();
-    bool input_fully_transparent();
-    const std::vector<fix15_short_t> factors;
-    const int radius;
-    chan_t** input_full;
-    chan_t** input_vert;
-    chan_t output[N][N];
-};
-
-void blur(
-    int radius, // Radius to grow (if > 0) or shrink (if < 0)
-    PyObject* blurred, // Dictionary holding the result of the operation
-    PyObject* tiles, // Input tiles, NxNx1 uint16 numpy arrays
-    PyObject* strands // Strands of contiguous tile coordinates
-    );
-
-// Gapclosing fill data utilities
-
-#ifdef SWIG
-%ignore DistanceBucket::distance;
-%ignore DistanceBucket::input;
-#endif
-
-// Distance data bucket for gap closing
-class DistanceBucket
-{
-  public:
-    explicit DistanceBucket(int distance);
-    ~DistanceBucket();
-    const int distance;
-    chan_t** input;
-};
-
-// Search the given nine-grid of flooded alpha tiles for
-// gaps up to a certain length, defined by the DistanceBucket,
-// writing the lengths found to the given distance tile
-// Returns true if any gaps were found.
-bool find_gaps(
-    DistanceBucket& bucket, PyObject* gap_output, PyObject* src_mid,
-    PyObject* src_n, PyObject* src_e, PyObject* src_s, PyObject* src_w,
-    PyObject* src_ne, PyObject* src_se, PyObject* src_sw, PyObject* src_nw);
-
-#endif
+#endif //MORPHOLOGY_HPP
