+++ lib/mypaintlib.hpp
@@ -23,7 +23,10 @@
 #include "colorchanger_crossed_bowl.hpp"
 #include "gdkpixbuf2numpy.hpp"
 #include "fastpng.hpp"
+#include "fill_constants.hpp"
 #include "fill_common.hpp"
 #include "floodfill.hpp"
+#include "gap_detection.hpp"
+#include "blur.hpp"
 #include "morphology.hpp"
 #include "brushsettings.hpp"
