+++ setup.py
@@ -579,7 +579,10 @@ def get_ext_modules():
         [
             'lib/mypaintlib.i',
             'lib/fill_common.cpp',
+            'lib/fill_constants.cpp',
             'lib/floodfill.cpp',
+            'lib/gap_detection.cpp',
+            'lib/blur.cpp',
             'lib/morphology.cpp',
             'lib/gdkpixbuf2numpy.cpp',
             'lib/pixops.cpp',
